+++ lib/hexapdf/type/font_simple.rb
@@ -48,7 +48,7 @@ module HexaPDF
       define_field :LastChar, type: Integer
       define_field :Widths, type: Array
       define_field :FontDescriptor, type: :FontDescriptor, indirect: true
-      define_field :Encoding, type: [Symbol, Dictionary, Hash]
+      define_field :Encoding, type: [Dictionary, Symbol]
 
       # Returns the encoding object used for this font.
       #
@@ -60,7 +60,7 @@ module HexaPDF
             encoding = HexaPDF::Font::Encoding.for_name(val)
             encoding = encoding_from_font if encoding.nil?
             encoding
-          when HexaPDF::Dictionary, Hash
+          when HexaPDF::Dictionary
             encoding = val[:BaseEncoding] && HexaPDF::Font::Encoding.for_name(val[:BaseEncoding])
             encoding ||= if embedded? || symbolic?
                            encoding_from_font